# docs/api-spec.yml
...
openapi: 3.1.0
info:
  title: Archivo Desclasificado — Editorial Service
  version: 0.1.0
paths:
...
  /videos:
    get:
      tags:
      - video-management
      summary: List All Videos
      description: 'Every generated video, newest first, excluding deleted ones.


        Only `status` filters server-side; platform, language and date are the

        caller''s to apply.'
      operationId: list_all_videos_videos_get
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/VideoGenerationStatus'
          - type: 'null'
          description: Filter by generation status.
          title: Status
        description: Filter by generation status.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VideoListItemResponse'
                title: Response List All Videos Videos Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /videos/stats:
    get:
      tags:
      - video-management
      summary: Video Stats
      description: 'Storage usage, counts by status, largest videos, and disk capacity.


        Counts come from the database and sizes from disk, so `missing_on_disk`

        reports rows marked generated whose file has vanished. Disk capacity is

        null, never 0, when it cannot be read.'
      operationId: video_stats_videos_stats_get
      parameters:
      - name: largest_limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 10
          title: Largest Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoStatsResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /videos/{video_id}/file:
    get:
      tags:
      - video-management
      summary: Download Video File
      description: 'Serve the MP4 itself.


        The admin panel cannot play `video_file_path` directly — it is a path on

        the server''s filesystem, not a URL. Starlette''s FileResponse honours

        Range requests, so the player can seek without pulling the whole file.'
      operationId: download_video_file_videos__video_id__file_get
      parameters:
      - name: video_id
        in: path
        required: true
        schema:
          type: string
          title: Video Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /videos/{video_id}:
    delete:
      tags:
      - video-management
      summary: Delete One Video
      description: 'Delete a video''s files and mark the record deleted.


        Soft delete: the row is retained so the retry lineage the audit trail

        reports stays intact. Shared narration and canonical subtitles are never

        removed, and a per-platform file is kept when another live record still

        points at it — the response says which and why.'
      operationId: delete_one_video_videos__video_id__delete
      parameters:
      - name: video_id
        in: path
        required: true
        schema:
          type: string
          title: Video Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoDeletionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /chapters/{chapter_id}/audit:
    get:
      tags:
      - video-management
      summary: Chapter Audit Trail
      description: 'Every generation attempt for a chapter, oldest first.


        Includes deleted attempts and retry lineage. `triggered_by` is always

        null: this service has no authentication, so no caller identity exists.'
      operationId: chapter_audit_trail_chapters__chapter_id__audit_get
      parameters:
      - name: chapter_id
        in: path
        required: true
        schema:
          type: string
          title: Chapter Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AuditEntryResponse'
                title: Response Chapter Audit Trail Chapters  Chapter Id  Audit Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AuditEntryResponse:
      properties:
        video_generation_id:
          type: string
          title: Video Generation Id
        platform:
          type: string
          title: Platform
        language:
          type: string
          title: Language
        status:
          type: string
          title: Status
        triggered_by:
          anyOf:
          - type: string
          - type: 'null'
          title: Triggered By
        triggered_at:
          type: string
          format: date-time
          title: Triggered At
        completed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Completed At
        deleted_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Deleted At
        retry_of_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Retry Of Id
        error_message:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Message
      type: object
      required:
      - video_generation_id
      - platform
      - language
      - status
      - triggered_by
      - triggered_at
      - completed_at
      - deleted_at
      - retry_of_id
      - error_message
      title: AuditEntryResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    LargestVideoResponse:
      properties:
        id:
          type: string
          title: Id
        chapter_id:
          type: string
          title: Chapter Id
        platform:
          type: string
          title: Platform
        language:
          type: string
          title: Language
        size_mb:
          type: number
          title: Size Mb
        video_file_path:
          type: string
          title: Video File Path
      type: object
      required:
      - id
      - chapter_id
      - platform
      - language
      - size_mb
      - video_file_path
      title: LargestVideoResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    VideoDeletionResponse:
      properties:
        video_generation_id:
          type: string
          title: Video Generation Id
        video_file_deleted:
          type: boolean
          title: Video File Deleted
        subtitle_file_deleted:
          type: boolean
          title: Subtitle File Deleted
        video_file_retained_reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Video File Retained Reason
        subtitle_file_retained_reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Subtitle File Retained Reason
      type: object
      required:
      - video_generation_id
      - video_file_deleted
      - subtitle_file_deleted
      - video_file_retained_reason
      - subtitle_file_retained_reason
      title: VideoDeletionResponse
...
    VideoGenerationStatus:
      type: string
      enum:
      - pending
      - generated
      - failed
      title: VideoGenerationStatus
      description: Lifecycle of a VideoGeneration record.
    VideoListItemResponse:
      properties:
        id:
          type: string
          title: Id
        chapter_id:
          type: string
          title: Chapter Id
        chapter_title:
          type: string
          title: Chapter Title
        platform_version_id:
          type: string
          title: Platform Version Id
        platform:
          type: string
          title: Platform
        language:
          type: string
          title: Language
        status:
          type: string
          title: Status
        video_file_path:
          anyOf:
          - type: string
          - type: 'null'
          title: Video File Path
        subtitle_file_path:
          anyOf:
          - type: string
          - type: 'null'
          title: Subtitle File Path
        size_mb:
          anyOf:
          - type: number
          - type: 'null'
          title: Size Mb
        error_message:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Message
        retry_of_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Retry Of Id
        created_at:
          type: string
          format: date-time
          title: Created At
        generated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Generated At
      type: object
      required:
      - id
      - chapter_id
      - chapter_title
      - platform_version_id
      - platform
      - language
      - status
      - video_file_path
      - subtitle_file_path
      - size_mb
      - error_message
      - retry_of_id
      - created_at
      - generated_at
      title: VideoListItemResponse
...
    VideoStatsResponse:
      properties:
        total_storage_mb:
          type: number
          title: Total Storage Mb
        count_by_status:
          additionalProperties:
            type: integer
          type: object
          title: Count By Status
        largest_videos:
          items:
            $ref: '#/components/schemas/LargestVideoResponse'
          type: array
          title: Largest Videos
        missing_on_disk:
          type: integer
          title: Missing On Disk
        disk_free_mb:
          anyOf:
          - type: number
          - type: 'null'
          title: Disk Free Mb
        disk_total_mb:
          anyOf:
          - type: number
          - type: 'null'
          title: Disk Total Mb
      type: object
      required:
      - total_storage_mb
      - count_by_status
      - largest_videos
      - missing_on_disk
      - disk_free_mb
      - disk_total_mb
      title: VideoStatsResponse
servers:
- url: http://localhost:8000
  description: Local development server